requirements.txt
torch>=1.7
numpy<1.21  # numba requires numpy<1.21,>=1.17
opencv-python~=4.5.5.62
torchvision~=0.11.2
scipy~=1.7.3
tqdm~=4.62.3
basicsr>=1.3.4.0
facexlib>=0.2.0.3
lmdb
pyyaml
tb-nightly
yapf
argparse~=1.4.0
Pillow~=9.0.0
gfpgan~=0.2.4
realesrgan~=0.2.3.0
watchdog==2.1.6

dlib~=19.23.0
matplotlib~=3.5.1